[Vertex AI] Renamed `RPCError` to `BackendError`
